moved all completion related functions to completions
